test(agent): isolate compaction extension coverage
Problem:
The compaction changes added assertions to an already oversized extension test file, which exceeded its repolint carry-forward budget and failed the PR lint job.

Root cause:
The new summary-input mode coverage was placed beside legacy extension cases instead of in a focused test file.

Fix:
Move the two compaction summary-input extension tests into a dedicated file without changing their behavior or widening the lint baseline.

Verification:
- go test ./internal/agent -run TestCompactionPrepareReplace
- go run ./tools/repolint
- git diff --check
